+++ Braintree/Configuration.cs
@@ -2,11 +2,14 @@
 
 using System;
 using System.Collections.Generic;
+using System.Net;
 using System.Text;
 using Braintree.Exceptions;
 
 namespace Braintree
 {
+    public delegate HttpWebRequest HttpWebRequestFactoryDelegate(string requestUriString);
+
     public class Configuration
     {
         public Environment Environment { get; set; }
@@ -17,6 +20,7 @@ namespace Braintree
         public string PrivateKey { get; set; }
         public string Proxy { get; set; }
         public string PublicKey { get; set; }
+        public HttpWebRequestFactoryDelegate HttpWebRequestFactory { get; set; }
 
         private int timeout;
         public int Timeout
@@ -25,8 +29,12 @@ namespace Braintree
             set { timeout = value; }
         }
 
-        public Configuration() {}
-        public Configuration(string accessToken)
+        public Configuration()
+        {
+            HttpWebRequestFactory = requestUriString => WebRequest.Create(requestUriString) as HttpWebRequest;
+        }
+
+        public Configuration(string accessToken) : this()
         {
             CredentialsParser parser = new CredentialsParser(accessToken);
             MerchantId = parser.MerchantId;
@@ -34,7 +42,7 @@ namespace Braintree
             Environment = parser.Environment;
         }
 
-        public Configuration(string clientId, string clientSecret)
+        public Configuration(string clientId, string clientSecret) : this()
         {
             CredentialsParser parser = new CredentialsParser(clientId, clientSecret);
             ClientId = parser.ClientId;
@@ -42,7 +50,7 @@ namespace Braintree
             Environment = parser.Environment;
         }
 
-        public Configuration(Environment environment, string merchantId, string publicKey, string privateKey)
+        public Configuration(Environment environment, string merchantId, string publicKey, string privateKey) : this()
         {
             if (environment == null) {
                 throw new ConfigurationException("Configuration.environment needs to be set");