+++ Braintree/ITransactionGateway.cs
@@ -27,6 +27,7 @@ namespace Braintree
         Result<Transaction> SubmitForSettlement(string id);
         Result<Transaction> SubmitForSettlement(string id, decimal amount);
         Result<Transaction> SubmitForSettlement(string id, TransactionRequest request);
+        Result<Transaction> UpdateDetails(string id, TransactionRequest request);
         string TransparentRedirectURLForCreate();
         Result<Transaction> Void(string id);
     }