+++ Braintree/TransactionGateway.cs
@@ -140,6 +140,13 @@ namespace Braintree
             return new ResultImpl<Transaction>(new NodeWrapper(response), gateway);
         }
 
+        public virtual Result<Transaction> UpdateDetails(string id, TransactionRequest request)
+        {
+            XmlNode response = service.Put(service.MerchantPath() + "/transactions/" + id + "/update_details", request);
+
+            return new ResultImpl<Transaction>(new NodeWrapper(response), gateway);
+        }
+
         public virtual Result<Transaction> SubmitForPartialSettlement(string id, decimal amount)
         {
             var request = new TransactionRequest();